decomposed our scopes into more maintainable files
